Added coverage for integration tests
