+++ tests-unit/test_client.py
@@ -1,0 +1,24 @@
+import os
+import sys
+import tempfile
+
+import pytest
+
+sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
+
+from app import create_app
+from db import init_db
+
+
+@pytest.fixture
+def client():
+    db_fd, db_path = tempfile.mkstemp()
+    app = create_app({'TESTING': True, 'DATABASE': db_path}, testing=True)
+
+    with app.test_client() as client:
+        with app.app_context():
+            init_db()
+        yield client
+
+    os.close(db_fd)
+    os.unlink(db_path)
